refactor(fpss): co-locate leftover mod.rs tests with their code
Move the FpssEvent / FpssControl variant-shape tests from mod.rs into
events.rs where those enums now live, and the ReconnectPolicy tests into
session.rs where the reconnect lifecycle lives. mod.rs now only holds
FpssClient, its impl, its Drop, and the module declarations.
